preentrega 1

+++ src/components/CartManager.js
@@ -42,14 +42,23 @@ class CartManager {
 
   //  Agregar un producto al carrito
   addProductToCart = async (cartId, productId) => {
-    const data = await this.getCarts();
-    const cart = data.find((cart1) => cart1.id === cartId);
+    let data = await this.getCarts();
+    const cart = data.find((cart1) => cart1.id === parseInt(cartId));
     if (cart) {
-      data = data.filter((carts) => carts.id != cartId);
-      cart.products.push({ product: productId, quantity: 1 });
-      data.push(cart);
-      await fs.writeFile(this.path, JSON.stringify(data, null, "\t"));
-      return "Se agrego correctamente al carrito";
+      let productExist = cart.products.find(
+        (product) => product.product === parseInt(productId)
+      );
+      if (productExist) {
+        productExist.quantity++;
+        await fs.writeFile(this.path, JSON.stringify(data, null, "\t"));
+        return "Se agrego correctamente al carrito";
+      } else {
+        data = data.filter((carts) => carts.id != parseInt(cartId));
+        cart.products.push({ product: parseInt(productId), quantity: 1 });
+        data.push(cart);
+        await fs.writeFile(this.path, JSON.stringify(data, null, "\t"));
+        return "Se agrego correctamente al carrito";
+      }
     } else {
       return "El carrito no existe";
     }
